pyproject.toml
[project]
name = "prism-council"
version = "0.1.0"
description = "PRISM Council - Decision-grade multi-model reasoning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
]
